Replace Python script with Parcel package manager
The main advantage is that it doesn't need any configuration at all.

An added benefit of Parcel solution is that it generates map files and so source files can be debugged under the navigator.

The Python script concatenated all js files => global variables are not accessible anymore => export and import shared variables.

Correct reference to DKG example which was detected by Parcel (doubled equal sign in <a href=="...">)

Problems in encrypt and generate_poly functions exported by lib.rs:
- They used u64 parameters and Python version called them directly,
whereas Parcel version generates intermediate wrapper functions
to convert to BigInt (in __exports.encrypt and __exports.generate_poly).
Unfortunately this conversion fails with Parcel
=> change their parameters to u32, which is the right thing
to do because this is the type of values that are really passed in
(generated by window.crypto.getRandomValues that fills a Uint32Array)
- Formula `seed1 << 32 + seed2` wasn't right because precedence of + operator
is greater than << operator. The formula has been corrected and put in a common function.
A Rust test has also been added to test this function.