Replace Python script with Parcel bundler
The main advantage is that it doesn't need any configuration at all.

An added benefit of Parcel solution is that it generates map files and so source files can be debugged under the navigator.

The Python script concatenated all js files => global variables are not accessible anymore => export and import shared variables.